# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VoteVoteType = (typeof VoteVoteType)[keyof typeof VoteVoteType];

export const VoteVoteType = {
  up: "up",
  neutral: "neutral",
  down: "down",
} as const;

export interface Vote {
  id: number;
  voter: string;
  dayNumber: number;
  activityIndex: number;
  voteType: VoteVoteType;
  createdAt: string;
  updatedAt: string;
}

export type VoteInputVoteType =
  (typeof VoteInputVoteType)[keyof typeof VoteInputVoteType];

export const VoteInputVoteType = {
  up: "up",
  neutral: "neutral",
  down: "down",
} as const;

export interface VoteInput {
  voteType: VoteInputVoteType;
}

export interface CustomActivity {
  id: number;
  dayNumber: number;
  proposedBy: string;
  name: string;
  description?: string;
  createdAt: string;
}

export interface NewCustomActivity {
  dayNumber: number;
  proposedBy: string;
  name: string;
  description?: string;
}

export interface ActivityComment {
  id: number;
  author: string;
  dayNumber: number;
  activityIndex?: number | null;
  customActivityId?: number | null;
  content: string;
  createdAt: string;
}

export interface NewComment {
  author: string;
  dayNumber: number;
  activityIndex?: number;
  customActivityId?: number;
  content: string;
}

export type CustomActivityVoteVoteType =
  (typeof CustomActivityVoteVoteType)[keyof typeof CustomActivityVoteVoteType];

export const CustomActivityVoteVoteType = {
  up: "up",
  neutral: "neutral",
  down: "down",
} as const;

export interface CustomActivityVote {
  id: number;
  voter: string;
  customActivityId: number;
  voteType: CustomActivityVoteVoteType;
  createdAt: string;
  updatedAt: string;
}

export interface ActivityMove {
  id: number;
  dayNumber: number;
  activityIndex: number;
  targetDayNumber: number;
  movedBy: string;
  createdAt: string;
}

export interface ActivityMoveInput {
  targetDayNumber: number;
  movedBy: string;
}

export interface Reservation {
  id: number;
  dayNumber: number;
  activityIndex: number;
  required: boolean;
  deadline?: string | null;
  updatedBy: string;
  updatedAt: string;
}

export interface ReservationInput {
  required: boolean;
  deadline?: string | null;
  updatedBy: string;
}