# lib/api-zod/src/generated/types/customActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomActivity {
  id: number;
  dayNumber: number;
  proposedBy: string;
  name: string;
  description?: string;
  createdAt: string;
}