# lib/api-zod/src/generated/types/customActivityVote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomActivityVoteVoteType } from "./customActivityVoteVoteType";

export interface CustomActivityVote {
  id: number;
  voter: string;
  customActivityId: number;
  voteType: CustomActivityVoteVoteType;
  createdAt: string;
  updatedAt: string;
}